+++ packages/wujie-core/src/common.ts
@@ -5,6 +5,8 @@ export interface SandboxCache {
   options?: cacheOptions;
 }
 
+export type appAddEventListenerOptions = AddEventListenerOptions & { targetWindow?: Window };
+
 // 全部无界实例和配置存储map
 export const idToSandboxCacheMap = window.__POWERED_BY_WUJIE__
   ? window.__WUJIE.inject.idToSandboxMap
@@ -39,7 +41,14 @@ export function addSandboxCacheWithOptions(id: string, options: cacheOptions): v
 // 分类document上需要处理的属性，不同类型会进入不同的处理逻辑
 export const documentProxyProperties = {
   // 降级场景下需要本地特殊处理的属性
-  modifyLocalProperties: ["createElement", "createTextNode", "documentURI", "URL", "getElementsByTagName"],
+  modifyLocalProperties: [
+    "createElement",
+    "createTextNode",
+    "documentURI",
+    "URL",
+    "getElementsByTagName",
+    "getElementById",
+  ],
 
   // 子应用需要手动修正的属性方法
   modifyProperties: [
@@ -109,6 +118,7 @@ export const documentProxyProperties = {
   // 需要从主应用document中获取的方法
   documentMethods: [
     "execCommand",
+    "caretPositionFromPoint",
     "createRange",
     "exitFullscreen",
     "exitPictureInPicture",
@@ -147,6 +157,9 @@ export const mainDocumentAddEventListenerEvents = [
   "selectionchange",
   "visibilitychange",
   "wheel",
+  "keydown",
+  "keypress",
+  "keyup",
 ];
 
 // 需要同时挂载到主应用document和shadow上的事件（互斥）
@@ -160,10 +173,13 @@ export const appWindowAddEventListenerEvents = [
   "load",
   "beforeunload",
   "unload",
+  "message",
+  "error",
+  "unhandledrejection",
 ];
 
 // 子应用window.onXXX需要挂载到iframe沙箱上的事件
-export const appWindowOnEvent = ["onload", "onbeforeunload", "onunload"];
+export const appWindowOnEvent = ["onload", "onbeforeunload", "onunload", "onerror", "onunhandledrejection"];
 
 // 相对路径问题元素的tag和attr的map
 export const relativeElementTagAttrMap = {
@@ -181,6 +197,7 @@ export const windowRegWhiteList = [
   /resizeObserver$|mutationObserver$|intersectionObserver$/i,
   /height$|width$|left$/i,
   /^screen/i,
+  /CSSStyleSheet$/i,
   /X$|Y$/,
 ];
 
@@ -188,10 +205,13 @@ export const windowRegWhiteList = [
 // 子应用的Document.prototype已经被改写了
 export const rawElementAppendChild = HTMLElement.prototype.appendChild;
 export const rawElementRemoveChild = HTMLElement.prototype.removeChild;
+export const rawElementContains = HTMLElement.prototype.contains;
 export const rawHeadInsertBefore = HTMLHeadElement.prototype.insertBefore;
 export const rawBodyInsertBefore = HTMLBodyElement.prototype.insertBefore;
-export const rawAddEventListener = EventTarget.prototype.addEventListener;
-export const rawRemoveEventListener = EventTarget.prototype.removeEventListener;
+export const rawAddEventListener = Node.prototype.addEventListener;
+export const rawRemoveEventListener = Node.prototype.removeEventListener;
+export const rawWindowAddEventListener = window.addEventListener;
+export const rawWindowRemoveEventListener = window.removeEventListener;
 export const rawAppendChild = Node.prototype.appendChild;
 export const rawDocumentQuerySelector = window.__POWERED_BY_WUJIE__
   ? window.__WUJIE_RAW_DOCUMENT_QUERY_SELECTOR__